+++ Modules/Tooltip.lua
@@ -5,8 +5,13 @@ function MS:FetchPlayerLockouts()
     local RaidLockouts = {}
     local DungeonLockouts = {}
     for i = 1, GetNumSavedInstances() do
-        local Name, _, _, _, IsLocked, _, _, IsRaid, _, DifficultyName, MaxEncounters, CurrentProgress, _, _ = GetSavedInstanceInfo(i)
-        local LockoutString = string.format("%s: %d/%d %s", Name, CurrentProgress, MaxEncounters, DifficultyName)
+        local Name, _, Reset, _, IsLocked, _, _, IsRaid, _, DifficultyName, MaxEncounters, CurrentProgress, _, _ = GetSavedInstanceInfo(i)
+        -- Convert Seconds to Days / Hours:Mins
+        local Days = math.floor(Reset / 86400)
+        local Hours = math.floor((Reset % 86400) / 3600)
+        local Mins = math.floor((Reset % 3600) / 60)
+        Reset = Days > 0 and string.format("%dd %dh %dm", Days, Hours, Mins) or string.format("%dh %dm", Hours, Mins)
+        local LockoutString = string.format("%s: %d/%d %s [%sReset|r: %s]", Name, CurrentProgress, MaxEncounters, DifficultyName, MS.AccentColour, Reset)
         if IsLocked then
             if IsRaid then
                 table.insert(RaidLockouts, LockoutString)
@@ -28,7 +33,7 @@ function MS:FetchPlayerLockouts()
         GameTooltip:AddLine("Raid |cFFFFFFFFLockouts|r", MS.DB.global.AccentColourR, MS.DB.global.AccentColourG, MS.DB.global.AccentColourB, 1)
         for _, Lockout in pairs(RaidLockouts) do
             local RaidTitle, RaidLockout = Lockout:match("([^:]+): (.+)")
-            local RaidTitle = MS.AbbrRaidTitles[RaidTitle:match("([^:]+)")] or RaidTitle
+            local RaidTitle = MS.AbbrInstances[RaidTitle:match("([^:]+)")] or RaidTitle
             local RaidDisplayString = MS.AccentColour .. RaidTitle .. "|r: " .. RaidLockout
             GameTooltip:AddLine(RaidDisplayString, 1, 1, 1, 1)
         end